Move reusable code into Easing::setupShaderEasing
