+++ js/effects/particle-displace.js
@@ -72,24 +72,7 @@ export default class ParticleDisplaceEffect extends Effect {
     }
     angle = (angle + 2 * Math.PI) % (2 * Math.PI);
     const distance = instance.config.distance || 0.5;
-    const easeInTime = Math.min(instance.config.easeInTime || 1000, instance.getPeriod() / 2);
-    const easeOutTime = Math.min(instance.config.easeOutTime || 1000, instance.getPeriod() - easeInTime);
-    // starts at 0, goes down to 1
-    const easeInProgress = uniforms.addUniform('easeInProgress', 'float', (ctx, props) => {
-      const time = fract((props.clock.getTime() - instance.timeBegin) / instance.getPeriod());
-      return Math.min(1, time / (easeInTime / instance.getPeriod()));
-    });
-    // starts at 1, goes down to 0
-    const easeOutProgress = uniforms.addUniform('easeOutProgress', 'float', (ctx, props) => {
-      const time = fract((props.clock.getTime() - instance.timeBegin) / instance.getPeriod());
-      return Math.min(1, (1 - time) / (easeOutTime / instance.getPeriod()));
-    });
-    const easeFuncs = {
-      none: '1.',
-      sine: `(1. - cos(PI * min(${easeInProgress}, ${easeOutProgress}))) / 2.`,
-      linear: `min(${easeInProgress}, ${easeOutProgress})`
-    };
-    const easeFunc = easeFuncs[instance.config.easeFunc || 'sine'];
+    const easeFunc = Ease.setupShaderEasing(instance, uniforms);
     vertexShader.mainBody += `
       vec2 offset;
       offset.y = cos(float(${angle}));
